Gauge: Send system information over a TcpStream
