+++ gauge/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+crossbeam = "0.8.2"
 serde = { version="1.0.159", features = ["derive"] }
 serde_json = "1.0.95"
-sysinfo = "0.28.4"
 systemstat = "0.2.3"
